Fixed links after moving content for future growth.
