+++ courses/one-day-overview/README.md
@@ -12,75 +12,75 @@ For a great talk/video that reinforces many of these ideas, see Andrew Gerrand's
 
 Easy to learn and reason about, trivial to machine-process.
 
-[Variables, Builtins, and Type Safety](http://play.golang.org/p/6w6hBNE75a) ([Source](../../topics/variables/example1/example1.go))
+[Variables, Builtins, and Type Safety](https://play.golang.org/p/B5mjJKPYLh) ([Source](../../topics/language/variables/example1/example1.go))
 
 [Constants](http://play.golang.org/p/ZHfzj2_Rse) ([Source](misc/consts/main.go))
 
-[Structs](http://play.golang.org/p/TAX6NpPaEu) ([Source](../../topics/struct_types/example1/example1.go))
+[Structs](https://play.golang.org/p/TEmOrIxl_P) ([Source](../../topics/language/struct_types/example1/example1.go))
 
 ## Pointers and Reference Types
 
 Safe and simple. Strings and slices solve many traditional pointer use-cases.
 
-[Arrays](http://play.golang.org/p/2D24t6fbW_) ([Source](../../topics/arrays/example1/example1.go))
+[Arrays](https://play.golang.org/p/wUzREuHhLY) ([Source](../../topics/language/arrays/example1/example1.go))
 
-[Slices](http://play.golang.org/p/AFb1SZ_1WZ) ([Source](../../topics/slices/example3/example3.go))
+[Slices](https://play.golang.org/p/Okc2EZG5_M) ([Source](../../topics/language/slices/example3/example3.go))
 
-[Appending](http://play.golang.org/p/BSNAUj2pd-) ([Source](../../topics/slices/exercises/exercise1/exercise1.go))
+[Appending](https://play.golang.org/p/3WKISOXA-L) ([Source](../../topics/language/slices/exercises/exercise1/exercise1.go))
 
-[Strings](http://play.golang.org/p/W3c_iWsvqj) ([Source](../../topics/slices/example5/example5.go))
+[Strings](https://play.golang.org/p/x0Q5ByzxGS) ([Source](../../topics/language/slices/example6/example6.go))
 
-[Variadic Functions](http://play.golang.org/p/5uDVuormwB) ([Source](../../topics/slices/example6/example6.go))
+[Variadic Functions](https://play.golang.org/p/aTGRT1rhoO) ([Source](../../topics/language/slices/example7/example7.go))
 
-[Maps](http://play.golang.org/p/FcY_0ckwOZ) ([Source](../../topics/maps/example3/example3.go))
+[Maps](https://play.golang.org/p/B2klwmqmPZ) ([Source](../../topics/language/maps/example2/example2.go))
 
-[Pointers](http://play.golang.org/p/FWmGnVUDoA) ([Source](../../topics/pointers/example2/example2.go))
+[Pointers](https://play.golang.org/p/6GUcA7-x3j) ([Source](../../topics/language/pointers/example2/example2.go))
 
-[Peano Pointers](http://play.golang.org/p/7XdrgbTfZn) ([Source](https://golang.org/doc/play/peano.go)) *
+[Peano Pointers](https://play.golang.org/p/7XdrgbTfZn) ([Source](https://golang.org/doc/play/peano.go)) *
 
 ## Methods and Interfaces
 
 Allow your code to grow gracefully without you.
 
-[Methods](http://play.golang.org/p/ovMH0wrl4B) ([Source](../../topics/methods/example1/example1.go))
+[Methods](https://play.golang.org/p/nxAwTRWk4N) ([Source](../../topics/language/methods/example1/example1.go))
 
-[Fibonacci Closures](http://play.golang.org/p/A0nH96VB4S) ([Source](https://golang.org/doc/play/fib.go)) *
+[Fibonacci Closures](https://play.golang.org/p/A0nH96VB4S) ([Source](https://golang.org/doc/play/fib.go)) *
 
-[First-Class Methods](http://play.golang.org/p/MNI1jR8Ets) ([Source](../../topics/methods/advanced/example1/example1.go))
+[First-Class Methods](https://play.golang.org/p/UP7qzHN-Au) ([Source](../../topics/language/methods/example3/example3.go))
 
-[Interfaces](http://play.golang.org/p/CaBE4Z8-VR) ([Source](../../topics/interfaces/exercises/exercise1/exercise1.go))
+[Interfaces](https://play.golang.org/p/06fecJbfE4) ([Source](../../topics/language/interfaces/exercises/exercise1/exercise1.go))
 
 ## Embedding and Composition
 
 An extension/wrapping mechanism based on composition, not inheritance.
 
-[Embedding](http://play.golang.org/p/wAV3xnKj60) ([Source](../../topics/embedding/example2/example2.go))
+[Embedding](https://play.golang.org/p/QncBd6A5A4) ([Source](../../topics/language/embedding/example2/example2.go))
 
-[Embedding and Interfaces](http://play.golang.org/p/_MiwwXZbVI) ([Source](../../topics/embedding/example3/example3.go))
+[Embedding and Interfaces](https://play.golang.org/p/vMEEJ7rOb4) ([Source](../../topics/language/embedding/example3/example3.go))
 
-[Composition](http://play.golang.org/p/QnkL-UIVJN) ([Source](../../topics/composition/example2/example2.go))
+[Composition](https://play.golang.org/p/ufFSFxCdEs) ([Source](../../topics/api/composition/example4/example4.go))
 
 ## Concurrency
 
 Goroutines and channels for convenient and safe concurrent algorithms.
 
-[Concurrent Pi](http://play.golang.org/p/RdbPXQcZHi) ([Source](https://golang.org/doc/play/pi.go)) *
+[Concurrent Pi](https://play.golang.org/p/RdbPXQcZHi) ([Source](https://golang.org/doc/play/pi.go)) *
 
 [Concurrent Prime Sieve](https://golang.org/s/prime-sieve) ([Source](https://golang.org/doc/play/sieve.go)) *
 
-[Select](http://play.golang.org/p/Sv_eWCWqiJ) ([Source](../../topics/channels/example4/example4.go))
+[Select](https://play.golang.org/p/TsJSagQawy) ([Source](../../topics/concurrency/channels/example4/example4.go))
 
 ## Standard Library
 
 Excellent value for its volume. I/O is especially powerful.
 
-[Packages](../../topics/exporting/example5/example5.go) and [Exporting](../../topics/exporting/example5/users/users.go)
+[Packages](../../topics/language/exporting/example5/example5.go) and [Exporting](../../topics/language/exporting/example5/users/users.go)
 
-[Errors](http://play.golang.org/p/PiSDQj1UCB) ([Source](../../topics/error_handling/example1/example1.go))
+[Errors](https://play.golang.org/p/aSjTxzNfP2) ([Source](../../topics/api/error_handling/example1/example1.go))
 
-[Error Variables](https://play.golang.org/p/FRnwmQx_ZI) ([Source](../../topics/error_handling/example2/example2.go))
+[Error Variables](https://play.golang.org/p/-vBG0m1Scs) ([Source](../../topics/api/error_handling/example2/example2.go))
 
-[Behavior As Error Context](http://play.golang.org/p/6GYqwSxHjI) ([Source](../../topics/error_handling/example4/example4.go))
+[Behavior As Error Context](https://play.golang.org/p/Aylgou6Gq0) ([Source](../../topics/api/error_handling/example4/example4.go))
 
 [Basic cURL](misc/curl1/gocurl.go)
