+++ topics/api/composition/example2/example2.go
@@ -38,10 +38,13 @@ type Storer interface {
 // =============================================================================
 
 // Xenia is a system we need to pull data from.
-type Xenia struct{}
+type Xenia struct {
+	Host    string
+	Timeout time.Duration
+}
 
 // Pull knows how to pull data out of Xenia.
-func (Xenia) Pull(d *Data) error {
+func (*Xenia) Pull(d *Data) error {
 	switch rand.Intn(10) {
 	case 1, 9:
 		return io.EOF
@@ -57,10 +60,13 @@ func (Xenia) Pull(d *Data) error {
 }
 
 // Pillar is a system we need to store data into.
-type Pillar struct{}
+type Pillar struct {
+	Host    string
+	Timeout time.Duration
+}
 
 // Store knows how to store data into Pillar.
-func (Pillar) Store(d *Data) error {
+func (*Pillar) Store(d *Data) error {
 	fmt.Println("Out:", d.Line)
 	return nil
 }
