+++ topics/api/composition/example3/example3.go
@@ -44,10 +44,13 @@ type PullStorer interface {
 // =============================================================================
 
 // Xenia is a system we need to pull data from.
-type Xenia struct{}
+type Xenia struct {
+	Host    string
+	Timeout time.Duration
+}
 
 // Pull knows how to pull data out of Xenia.
-func (Xenia) Pull(d *Data) error {
+func (*Xenia) Pull(d *Data) error {
 	switch rand.Intn(10) {
 	case 1, 9:
 		return io.EOF
@@ -63,10 +66,13 @@ func (Xenia) Pull(d *Data) error {
 }
 
 // Pillar is a system we need to store data into.
-type Pillar struct{}
+type Pillar struct {
+	Host    string
+	Timeout time.Duration
+}
 
 // Store knows how to store data into Pillar.
-func (Pillar) Store(d *Data) error {
+func (*Pillar) Store(d *Data) error {
 	fmt.Println("Out:", d.Line)
 	return nil
 }
