+++ topics/language/arrays/README.md
@@ -103,9 +103,9 @@ This is subject to be different in different processors. For this content, the f
 
 ## Code Review
 
-[Declare, initialize and iterate](example1/example1.go) ([Go Playground](http://play.golang.org/p/wUzREuHhLY))  
-[Different type arrays](example2/example2.go) ([Go Playground](http://play.golang.org/p/tyOZ5_zBUN))  
-[Contiguous memory allocations](example3/example3.go) ([Go Playground](http://play.golang.org/p/DyZ7spMgZ3))  
+[Declare, initialize and iterate](example1/example1.go) ([Go Playground](https://play.golang.org/p/wUzREuHhLY))  
+[Different type arrays](example2/example2.go) ([Go Playground](https://play.golang.org/p/tyOZ5_zBUN))  
+[Contiguous memory allocations](example3/example3.go) ([Go Playground](https://play.golang.org/p/DyZ7spMgZ3))  
 
 ## Exercises
 
@@ -113,7 +113,7 @@ This is subject to be different in different processors. For this content, the f
 
 Declare an array of 5 strings with each element initialized to its zero value. Declare a second array of 5 strings and initialize this array with literal string values. Assign the second array to the first and display the results of the first array. Display the string value and address of each element.
 
-[Template](exercises/template1/template1.go) ([Go Playground](http://play.golang.org/p/H1jTYxk7o6)) | 
-[Answer](exercises/exercise1/exercise1.go) ([Go Playground](http://play.golang.org/p/i_2oDZ1ZSg))
+[Template](exercises/template1/template1.go) ([Go Playground](https://play.golang.org/p/H1jTYxk7o6)) | 
+[Answer](exercises/exercise1/exercise1.go) ([Go Playground](https://play.golang.org/p/i_2oDZ1ZSg))
 ___
 All material is licensed under the [Apache License Version 2.0, January 2004](http://www.apache.org/licenses/LICENSE-2.0).
