+++ topics/language/exporting/example4/example4.go
@@ -8,7 +8,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/ardanlabs/gotraining/topics/exporting/example4/users"
+	"github.com/ardanlabs/gotraining/topics/language/exporting/example4/users"
 )
 
 func main() {
