+++ topics/language/methods/README.md
@@ -17,8 +17,8 @@ http://www.goinggo.net/2014/05/methods-interfaces-and-embedded-types.html
 ## Code Review
 
 [Declare and receiver behavior](example1/example1.go) ([Go Playground](https://play.golang.org/p/nxAwTRWk4N))  
-[Named typed methods](example2/example2.go) ([Go Playground](http://play.golang.org/p/9WeR1rShIa))  
-[Function/Method variables](example3/example3.go) ([Go Playground](http://play.golang.org/p/UP7qzHN-Au))  
+[Named typed methods](example2/example2.go) ([Go Playground](https://play.golang.org/p/9WeR1rShIa))  
+[Function/Method variables](example3/example3.go) ([Go Playground](https://play.golang.org/p/UP7qzHN-Au))  
 [Function Types](example4/example4.go) ([Go Playground](https://play.golang.org/p/II6bnF3WKj))
 
 ## Exercises
@@ -27,7 +27,7 @@ http://www.goinggo.net/2014/05/methods-interfaces-and-embedded-types.html
 
 Declare a struct that represents a baseball player. Include name, atBats and hits. Declare a method that calculates a players batting average. The formula is Hits / AtBats. Declare a slice of this type and initialize the slice with several players. Iterate over the slice displaying the players name and batting average.
 
-[Template](exercises/template1/template1.go) ([Go Playground](http://play.golang.org/p/IG5uqVRTrc)) | 
-[Answer](exercises/exercise1/exercise1.go) ([Go Playground](http://play.golang.org/p/NcpOSoI0rn))
+[Template](exercises/template1/template1.go) ([Go Playground](https://play.golang.org/p/IG5uqVRTrc)) | 
+[Answer](exercises/exercise1/exercise1.go) ([Go Playground](https://play.golang.org/p/NcpOSoI0rn))
 ___
 All material is licensed under the [Apache License Version 2.0, January 2004](http://www.apache.org/licenses/LICENSE-2.0).
