+++ topics/standard_library/http/api/models/users.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/ardanlabs/gotraining/topics/http/api/app"
+	"github.com/ardanlabs/gotraining/topics/standard_library/http/api/app"
 )
 
 // UserAddress contains information about a user's address.
